requirements.txt
flask==2.3.3
flask-cors==4.0.0
pandas==2.1.0
PyMuPDF==1.23.5
faiss-cpu==1.7.4
numpy==1.25.2
sentence-transformers==2.2.2
openai==0.28.1
langchain==0.0.267
